refactor: Remove old makeXMLString func
