+++ editor.planx.uk/src/@planx/components/Send/uniform/index.ts
@@ -5,7 +5,7 @@ import { Store } from "../../../../pages/FlowEditor/lib/store";
 import { getBOPSParams } from "../bops";
 import { findGeoJSON } from "../helpers";
 import { CSVData } from "../model";
-import { OLD_makeXmlString } from "./xml";
+import { UniformPayload } from "./UniformPayload/model";
 
 type UniformFile = {
   name: string;
@@ -51,12 +51,23 @@ export function getUniformParams(
 
   // this is the body we'll POST to the /uniform endpoint - the endpoint will handle file & .zip generation
   return {
-    xml: OLD_makeXmlString(passport, sessionId, uniqueFiles, hasBoundary),
+    xml: makeXmlString(passport, sessionId, uniqueFiles, hasBoundary),
     csv: makeCsvData(breadcrumbs, flow, passport, sessionId),
     geojson: geoJSONBoundary,
     files: uniqueFiles,
     sessionId,
   };
+}
+
+export function makeXmlString(
+  passport: Store.passport,
+  sessionId: string,
+  files: string[],
+  hasBoundary: boolean,
+): string {
+  const payload = new UniformPayload(sessionId, passport, files);
+  const xml = payload.buildXML();
+  return xml;
 }
 
 // create a CSV data structure based on the payload we send to BOPs
